Fix a dependency order issue.
The built js files didn't work because of a wrong dependency : a file
used a wrong module as its own. With grunt serve everything worked
because of luck, not with grunt build. To fix this issue once and for
all, I migrated everything to use the dependency injection of the
module.