SO-42: implement the refresh_token.
If the user is more than 1h on navigator, its access_token will expire
and not be replaced (even with a refresh_token).

This commit uses the refresh_token to replace the access_token when it
expires. Without refresh_token (the "remember me" wasn't checked) we
show a warning when the session expires.