Fix nvd3 update 1.8.1 -> 1.8.2.
The `~` version in bower updated the 1.8.1 version of nvd3 to 1.8.2
which changed "Lots of community bugfixes and a few extra
minor features" (yay semver...).

The tooltip positionning changed and started throwing
`TypeError: position is not a function`. The new version works without
tooltip.position so we don't need it anymore.