Improve application bootstrap.
While we can't easily break our application in two -- a module to
bootstrap the UI (the navbar, the white label configuration, the ajax
loader) and the rest -- we can show the ajax loader quite early in the
bootstrap process. This way, the user will see that our UI isn't broken
and is still loading.